Zero-extend the horizontal radius in vector_op_unk2
vector_op_unk2 (asmorig/seg006.asm:2552) widens polarRadius2D's 16-bit result
into a 32-bit value by writing a zero high word next to it:

        call    polarRadius2D
        add     sp, 4
        mov     [bp+var_4], ax      ; low word
        mov     [bp+var_2], 0       ; high word - explicitly zero

var_4 and var_2 are adjacent, and the pair is used as one long from then on,
both in the direct comparison against |y|

        mov     ax, word ptr [bp+var_y]
        mov     dx, word ptr [bp+var_y+2]
        cmp     [bp+var_2], dx
        jg      short loc_2646E
        jl      short loc_26435
        cmp     [bp+var_4], ax
        jnb     short loc_2646E

and in the sin80/cos80 branch, which pushes [var_2]:[var_4] into __aFlmul.

The C port assigned the int return straight into a long, which sign-extends.
polarRadius2D returns hypot of two 16-bit components, so it exceeds 0x7FFF
for radii of 32768 and up - well inside the range a track produces. Those
came out negative, temp >= y flipped, and the function returned 0x1E / 0x1F,
the near-vertical indices, instead of a compass direction. The result feeds
byte_4393D in transformed_shape_op (shape3d.c:250).

Cast through unsigned short so the widening matches.

+++ src/restunts/c/math.c
@@ -680,7 +680,9 @@ int vector_op_unk2(struct VECTOR* vec) {
 	
 	y = abs(vec->y);
 	
-	temp = polarRadius2D(abs(vec->x), abs(vec->z));
+	// The original widens the 16-bit radius with an explicit zero high word
+	// (mov [bp+var_4], ax / mov [bp+var_2], 0), not with a sign extension.
+	temp = (unsigned short)polarRadius2D(abs(vec->x), abs(vec->z));
 	
 	if (sin80 != cos80) {
 		//fatal_error("sin80 != cos80 - not observed");
